wayland/lockscreen_overlay: Fix state not being persisted across windows
For https://invent.kde.org/plasma/plasma-mobile/-/merge_requests/270

Currently, when a surface is allowed to be a lockscreen overlay, the
property is set on the window. However, when the window is recreated
(ex. QML window.visible = false, window.visible = true) and the surface
remains the same, this state is lost.

This MR architecturally is probably not the way to go but I just wanted to
get something to work off of for discussion. It adds the lockscreen
overlay property on SurfaceInterface, and ensures that Window's property
is up to date with the surface.