backends/drm: don't use UUID to identify outputs
The UUID is only set later by the output configuration system, so it can't be used in
the drm backend when deciding the dpms state of a "new" output.
This changes it to use EDID ID, EDID hash or connector name instead, which are valid
much earlier and are reliable enough for the dpms workaround.

BUG: 493879
BUG: 506135
BUG: 505953