scene/workspacescene: never put some but not all quickly updating surfaces on overlays
If we have to repaint the primary plane anyways, it's not going to provide an efficiency
or latency improvement - at least not with the current way we use them.
On drivers where atomic tests with multiple planes are slow, it may also cause us to
switch between different overlay candidates frequently, causing dropped frames.