Fix "activate and raise" action with panels
With the "activate and raise" action, the window manager should activate
and raise the window but not pass the click.

If a window is focusable, then only the mouse press should be filtered
out.

If a window is not focusable, there is some sophisticated code to see
whether the window is obstructed by any other window (because isActive()
is always false for such windows so we need another way to determine
whether it's the first click). If the window is obstructed, then no click
will be passed but the action will still be performed, in other words,
the window will be activated (noop) and raised. However, this breaks if
a window has transients, which is the case with panels. We want the click
to be passed along even though the window is covered by a child (it
doesn't matter whether it accepts focus).

Given that primarily only special surfaces don't accept focus, e.g.
panels, and they are expected to receive clicks, the special stacking
order code path can be replaced with a `return false;` statement. Which
is also identical to the MouseActivate case.

See also e8dad997fae98e3c0c40653667dd22a002747556

BUG: 461414