inputmethod: Ensure the input method stays sane if it gets unmapped
When the panel would unmap (i.e. QWindow::setVisible(false) on the
client) we'd end up with a keyboard in a broken state.
Instead, mark it as minimised so it can be detected as such and when we
try to show it again, instead of trying to show an unexistant window,
tell the input method that we are back in business.

Also includes a handy qml tool to assess the state of the virtual
keyboard.

Fixes https://invent.kde.org/plasma/plasma-keyboard/-/issues/13