inputmethod: Ensure the input method stays sane if it gets unmapped
When the panel would unmap (i.e. QWindow::setVisible(false) on the
client) we'd end up with a keyboard in a broken state. React to
unmapping by hiding as per usual.

When calling show then we don't want to try to show the unmapped window
so make sure we account for that case. Instead ensure the context is
properly set up for the input method.

Also includes a handy qml tool to assess the state of the virtual
keyboard.

Fixes https://invent.kde.org/plasma/plasma-keyboard/-/issues/13