Tidy code in Window::mousePressCommandConsumesEvent()
The "activate and lower" action is available only for inactive windows.
Given that, it can be handled the same way as "activate and raise" and
"activate" cases. And in general, there is no good reason why it needs
to be different from the other two. This makes the code a bit nicer, you
won't need to wonder why the MouserActivateAndLower case is special.