plugins/nightlight: Drop setLocation from the d-bus api
With knighttime, this method is no longer necessary. Technically, it is
an API breaking change. But, it is an internal API, it was meant to be
only used by the geolocation KDED module, which has been dropped recently.

It makes the d-bus api more polished, it no longer contains any internal
implementation details.