draft: underlay support
Instead of just allowing the compositor to pick an item (tree) out of the primary
plane and put it on a plane above the primary one, this also makes it possible to
take an individual item and put it on a plane below the primary one.
To ensure the item is visible, we draw a transparent hole where it would be on the
primary plane.
The result of that change is that we can put items on hardware planes even when
they're partially covered by something else above it. The biggest use case for this
is videos with logos or subtitles on top of them, for which you'd otherwise need
multiple overlay planes - that all draw additional power.

Draft because this modified ItemTreeView to be only about an individual item. It
should instead use its own view that's about exactly that concept.