screencast: Resizable Virtual Monitors
When streaming a virtual monitor to a remote server, we want the client
to be able to be in charge of the virtual monitor size.

Pipewire has a path built in that allows the client and server to
negotiate a size that matches their constraints.

Gnome uses this for their resizing of virtual monitors, so clients will
have the expectation that kwin will match. It also allows the virtual
monitor path via the xdg-desktop-portal to not need additional
sidechannels.