backends/drm: avoid dropping the color pipeline cache for empty pipelines
When testing new output configurations, we (currently) first set an empty pipeline
and only afterwards use the real one. Before this commit, that always dropped the
cache, which meant the next commit had to re-calculate LUTs and create a new drm
property blob, which also causes the kernel to re-program the LUT.
With this change, the same blob is re-used, and so changing brightness with HDR
or software brightness is much smoother.