add an autotest for the drm backend that runs on real KMS
This currently just does very basic modeset and presentation tests, using the
presentation time protocol to check if that worked correctly. More advanced
checks, like using a writeback connector or checking drm state directly, can be
added later on top.

This also contains test cases for direct scanout and overlay planes, but to make
that work, we'll have to add a test client that uses EGL or otherwise supports
dmabuf feedback. It's also still unclear how to make that work in CI with vkms,
maybe udmabuf is an option, but I haven't tested that yet.