plugins/magnifier: Check whether fbo is valid
We can succeed creating a texture but we can still potentially fail to
create the fbo.

This change has a behavior change though. If an fbo cannot be created,
the zoom factor will still be updated so kwin doesn't potentially end up
busy looping. This shouldn't happen in practice though.