a11ykeyboardmanager: Send second modifier press to screenreader
We have code so that when a grabbed modifier is pressed twice within
a short timeframe the key is processed normally.

However in this case we also want to let the screenreader know about
the key, even though it should not be grabbed.

Orca uses this to announce the change in capslock state when pressing
capslock twice.

BUG: 507545


(cherry picked from commit cc1d5a93cb329df4e249d6397e0c6ea280cb7cc6)

Co-authored-by: Nicolas Fella <nicolas.fella@gmx.de>