kcms/screenedges: move to Display & Monitor group
It currently lives in the Mouse & Touchpad group, which makes some
technical sense in that screen edges can only be easily interacted with
using a pointing device.

However this is fairly unintuitive as a grouping criterion. Given the
clearer connection to screens, people will probably expect to find it
under Display & Monitor. Move it there.