backends/libinput: Check if a device is virtual, and expose that to DBus
Virtual input devices are common for user-space graphics tablet driver
software, the most popular of which is probably OpenTabletDriver. (There
are known proprietary ones as well.) These play fine with KWin, but we
explicitly want to show the user in the KCM that these devices should
not be configured through KWin, and eventually we may want to do some
special "passthrough" for them.

Hence why we need a property to uniquely identify them. Which is quite
hard in fact, the only reliable way I found was to check their UDev
devpath of which uinput devices always start with /devices/virtual.