backends/libinput: Don't needlessly write default config values
Previously, the libinput device would write config values even if they
were defaults. So if you're like me and tend to mess with your
configuration a lot, there's a ton of useless lines in your config.

But saving kilobytes of disk space isn't the real win here, it's that
before it would be hard to change any default values and have it affect
existing users. Depending on how long the user was using their machine
and configuring their devices, it could be possible they have an
entirely duplicate (and possibly outdated) config sitting on their
computer.

This does not have an effect on any existing configurations, you must
write to the key again for it to be deleted.