placementtracker: special-case custom tiled windows
Unlike with quick tiling, windows in custom tiles get untiled when the output
they were on gets removed, so this needs special casing or the placement tracker
won't restore the window state after hotplugs.
This also enables the output changes autotest for this case, as it's now fixed,
and removes the (now wrong) autotest from the tiles test.

BUG: 501306
BUG: 500242