implement mirroring in a better way
Instead of awkwardly changing the scale and position of outputs to make their
viewports overlap, this makes use of the split between logical and backend
outputs and only creates one logical output for the two backend outputs.
This has the benefit that window management code, effects and applications only
see the one output, which removes a whole bunch of user visible problems.

To render a logical output on a display with potentially mismatching aspect
ratio, this also adds infrastructure to have an offset (black bar) when
compositing on a mirroring display. With the projection matrix and item
renderer adjusted to handle that offset, effects only need very few changes.

BUG: 481222