wayland: Store xdg session data in a KSharedDataCache
The KSharedDataCache has a couple of advantages:

- the session data will be memory mapped, kwin won't need extra open()
  calls when an application starts
- the KSharedDataCache can evict very old items, this can be useful to
  keep the session storage size at a reasonable level

On average, kwin needs about 250 bytes to store session data, which is
rounded up to 512 bytes. With a 5 MiB default cache size, the session
storage will be able to store about 10000 sessions, which is a fair
compromise between the default cache size and compositor memory usage.