wayland: Prevent restoring windows offscreen from xdg session data
The xdg session data contains absolute window position. The problem
with that is that the output layout may change and that absolute
position can end up somewhere offscreen. For example, if the associated
output gets disabled.

This change makes XdgToplevelWindow also remember the output layout
hash. It is not the only solution, there are other alternatives, but
it is a good starting point to prevent kwin from restoring invalid
state, other options will need more brainstorming.