wayland/colormanagement: add workaround + configuration for Windows HDR apps
Many Windows HDR games have their own configuration for display brightness.
That configuration however relies on the Windows behavior of copying HDR content
to the screen unmodified, without any reference luminance or HDR metadata.
As a middle ground between that terrible behavior and not having Windows games
work as expected, this adds a workaround for these applications:
If a Windows application is detected, its reference, max average and maximum
luminance are overwritten with values from KWin's config, which are written to
by KScreen's HDR calibration app.

This means that while Windows games will work as expected on the screen the
configuration was written for, they will still be tonemapped properly when on
a different screen, or when taking a screenshot or screencast.

"detecting" Windows applications is implemented in two different ways, one is
simply the explicit "windows_scrgb" image description from the Wayland protocol,
and the other is a more hacky check for the executable name.
Once it's possible to set a "windows_pq" image description as well, and both
gamescope and Wine use that, we can remove the more hacky approach again.

BUG: 509114