wayland/colormanagement: fix the absolute rendering intent
The ICC defines the absolute rendering intent as a transformation from relative
colorimetric to absolute colorimetric:

XYZ_abs = XYZ_rel * XYZ_mediawhite / XYZ_pcswhite

With display profiles however, full adaptation to the display white point is
assumed, so media white is always the same as pcs white, and the calculation
cancels out. Because of this, the absolute colorimetric rendering intent is
the exact same as the relative one.

As we internally use the absolute rendering intent with the assumption of
"no adaptation", this commit doesn't drop it, but merely renames it.