compositor: add support for underlays
There are many, many situations, in which we can't put surfaces on an overlay,
for example
- the window has server side decorations with rounded corners
- there's subtitles on top of it
- there's an entire window on top of it

These situations can still work however, often even more efficiently, by
putting the overlay plane below the primary plane and painting a transparent
hole on the primary plane where the surface would normally be. Other elements,
like window outlines, other windows and subsurfaces are then painted on top of
the transparent hole, and are thus still visible.