autotests: Add xwayland drag-and-drop tests
The drag-and-drop bridging code between X11 and Wayland applications is
difficult to follow, and it is somewhat fragile to work with.

This change adds some basic tests to verify how drag-and-drop works.
Unfortunately, it brings quite a lot of boilerplate code that implements
the XDND protocol and the XdndSelection selection.