x11window: synchronize mapping with window suspended state
This way, effects can force windows to be mapped, which keeps them updating even
if they're on a different virtual desktop from the current one, when they're
being recorded with a screencast, or when animating from one virtual desktop to
the other.

This also works around a wrongly painted frame when x11 windows get mapped, by
using a timer to avoid unmapping and then immediately mapping the window again
when an effect changes the visibility of the window to animate for example
minimizing or virtual desktop switching.