tests: Simulate real keyboard handling in testFakeEventFallback
The current test monitors a key event, then matches against received key
codes. This does not correctly simulate real client behaivour. We need
to track keymaps and modifiers. The handling of the capital "A" is
currently incorrect, as we need the shift to be set as a modifier, not
as a key.

This is groundwork for a pending patch improving this test and the
relevant code.