tests: Simulate full keyboard handling in testFakeEventFallback
The current test monitors a key event, then matches against received key
codes. This does not correctly simulate real client behaivour. We need
to track keymaps and modifiers.

A few things were wrong:
- The handling of the capital "A" is
currently incorrect, as we need the shift to be set as a modifier, not
as a key.
- QKeySequenceFromText does not extract all the letters from a given
string, it's for extracting things in the form "control+l" which isn't
what we have here
- It didn't handle UTF-16
- If a keysym doesn't exist in our current we need to switch the layout