Switch most uses of ColorDescription to a shared pointer
ColorDescription is a pretty big class, and makes lots of other classes
unnecessarily large by being in them; by putting it in a shared pointer, we
make some allocations in hot code paths more efficient.

This commit doesn't add any caching for color descriptions yet, there can still
be two descriptions that are different pointers but have the same contents.
Fixing that is a task for later.