Remove a SeatInterface::setDragTarget() overload
While it seems useful at first, it can be a footgun in situations where
notifyPointerMotion() or notifyTouchMotion() are reordered after
setDragTarget(). It is less error prone if you are always expected to
provide a position to setDragTarget().

Long term design goal is to provide enter(), leave(), and motion() hooks
for drag events, similar to exisiting notifyPointerEnter(),
notifyPointerLeave(), and notifyPointerMotion().