Fix dnd in Firefox bookmarks toolbar popups
When the pointer crosses between popups, kwin will send a motion event
for the old popup whose coordinates lie outside the corresponding surface
followed by a leave and an enter event for another popup. That motion
event confuses Firefox, it thinks that the pointer has left the popup
so it should be closed.

This change re-arranges the order of motion and enter/leave events so
the latter come first. This is still not a great solution, a motion
event will be sent even if an enter event has been sent, but it is good
enough for the time being.

BUG: 508775