compositor: allow reordering planes, to make underlays work on Intel
On AMD, the primary plane is at zpos=2, and you can put an overlay plane below
the primary one. However, on Intel this doesn't work, because the primary plane
is fixed at zpos=0, and the overlay planes at zpos=1,2,3 respectively.

To fix that, and to make the overlay plane logic more flexible, this changes
the compositor to include all planes and all content in the overlay plane
assignment.

Instead of the scene providing a list of overlays and underlays, it simply
provides a list of items to be put on layers - with the scene's container item
representing the scene that is to be composited. The compositor then attempts
to assign output layers to each item, and if that fails, it falls back to
either primary+cursor, or primary-only without direct scanout as safe choices.

This also unifies the direct scanout logic with overlay planes, as direct
scanout is now just an alternative way the scene can choose items to be put
on output layers.