core/colorpipeline: fix comparisons between identity pipelines
In the drm backend, we sometimes had comparisons between empty pipelines that
just had different input ranges. As that input range is really just an
implementation detail, this changes it to only compare the actual operations.
If there's no operations, the input range and input color space don't matter
anyways.