scripting: add invalidateSceneGraph() slot to WindowThumbnailItem to release scenegraph resources
Qt will call this slot on the rendering thread if a QQuickWindow is hidden or closed.

Not handling this slot leads to the situation when we can't queue cleanup via QQuickWindow::scheduleRenderJob()
because we have no quick window anymore. As a result ThumbnailTextureProvider and its texture is never released
leading to a GPU memory leak.

doc: https://doc.qt.io/qt-6/qquickitem.html#graphics-resource-handling