plugins/windowaperture: Rewrite it without relying on tainted object properties
Currently, if you press and hold the Meta+D shortcut, some windows can
become permanently translucent. The root cause of that is that

   w.offCornerId = 42;
   console.log(w.offCornerId);

may print "undefined". Ulf (a core QJSEngine/QQmlEngine developer)
confirmed that this can't work properly.

As an alternative solution, this change rewrite the window aperture
effect so it doesn't rely on tainted object properties. In long term, it
is probably for the better as it makes running all scripts in a single
engine easier.

Ulf said that this can't work properly.