Provide fake virtual desktop information to X clients
We would like to add support for per screen virtual desktops, however
the problem is then that the desktop layout needs to be reported to X
clients.

In practice, as we can tell, the virtual desktop information is not used
heavily. Neither GTK nor Qt use it.

This change puts the corresponding behind a guard. If the
KWIN_XWAYLAND_ENABLE_NETWM_DESKTOP=1 environment variable is not set,
kwin will report a 1x1 desktop grid, windows will be told that they are
on a single desktop.

The other alternative is not to set desktop related properties at all,
but there is a higher risk of breaking apps due to missing
_NET_DESKTOP_GEOMETRY, etc.