Make DragAndDropFilter end touch dnd
The SeatInterface tries to be helpful by handling some drag-and-drop
business, but in long term, it is not so a good idea. First, the
wl_data_device-specific code needs to be aware of the input device that
drives the dnd session, despite the fact that the relevant protocol bits
are agnostic about the input device. Some steps have already been taken
to clean up this in order to support tablet dnd. Second, it implies that
the coordinates in all pointing input device events should be in the
same coordinate space, i.e. the global coordinate space. But we want to
move away to a simpler design to make sub-surface-related code more
straightforward.

This change moves touch dnd termination logic to the dnd event filter in
order to make the architecture more consistent.

Note that kwin still keeps sending wl_touch.cancel and wl_touch.up
events during dnd to the client. This was not changed in this MR, but we
should probably consider doing that later.