inputmethod: Tie the input method  we run to the keyboard layout
The input method we are using is emminently a contextual process and is
tightly correlated to the language we are meaning to use. And so is the
keyboard layout.

Tie both of these settings so that we can get the best input method at
any given time.

This is accompanied by a change in the keyboards kcm in plasma-desktop.