autotests/integration: make the sticky keys test more complete
And fix modifier state caused by sticky keys not being reported to keystate
protocol users.
Sticky key locking still doesn't work properly with this, as we lock keys on
press, but libkxbcommon afterwards unlocks keys on release. The relevant checks
in the test are commented out until that's fixed.

CCBUG: 509714