wayland: don't take presentation time feedbacks from surfaces
Just because we take the current surface state into account when rendering a
frame, doesn't mean we actually present that frame. This can happen for example
with the following sequence:
1. KWin composites frame A, takes presentation feedback from surface
2. KWin composites frame B, because triple buffering.
   There's no presentation feedback on the surface anymore
3. The commit thread replaces frame A with frame B,
   because rendering went faster than expected

The client commit was still presented, just in a different frame, so the
presentation feedback must not be discarded.

(cherry picked from commit e1f60633dc06119b2ed3629d4cb1030b122dca80)