ButtonsRebindsFilter: Fix my bad ring math, allow completing a circle
When I initially wrote this code I noticed a strange bug when going
between the degrees of 0 and 360. The rebound button or scroll would
"glitch" and momentarily think it was going in the opposite direction.
Take a ring rebound to scrolling in Firefox for example, hit the bottom/
top of a page, and you could see (while still continuously circling your
ring) that it scrolls in the opposite direction for just a moment.

The reason why this happens is simple: circles! For example, going from
ring position 0 to 355 would previously calculate a delta of... negative
355 degrees. That's not how circles work, but we can make a safe
assumption with how ring events work: they are always a continuous
motion. We also don't know how big the device's increments are, but it's
safe to assume they are way less than a half-circle (180 degrees.)

So the solution is: if the delta is humungous big (>180 degrees) like in
our 0->355 case, then it's crossing over the 0/360 boundary and should
be corrected. This makes continuous, encircling ring movement feel way
better.