backends/drm: add support for writeback connectors
Writeback connectors are just like normal connectors, but instead of sending
the composition result to a screen, it gets written into a buffer that we can
provide - which is very useful for tests.
This adds the basic infrastructure to use writeback connectors for autotests,
and adds a test for night light on the drm backend that uses this to compare
the result with a purely CPU-side calculation of the result.