backends/drm: add support for writeback connectors in autotests
Writeback connectors are just like normal connectors, but instead of sending
the composition result to a screen, it gets written into a buffer that we can
provide.
This commit adds the infrastructure to use writeback connectors in tests, and
adds a test case that verifies the color accuracy of our night light
implementation.