fakeinput: Support arbitrary keysyms
When getting a keycode that cannot map to a given keycode in the current
keymap install a temporary keymap within kwin and for all clients that
contains the current symbol. This allows us to inject any unicode symbol
from fake input regardless of the current keymap.

This commit also improves the mapping of artbitrary levels to be more
technically correct.

This will also be relevant when libei gets keysym support.