ButtonsRebindsFilter: Remove assumption about how config keys are stored
When we get a notification from the config watcher, we filter it based
on the root key - because the kcminputrc file is also used for misc
libinput device configuration.

However the way we did that made an assumption about how they were
stored, specifically that the config groups were only one or two levels
deep. Well, I broke that assumption when adding ring support because I
nested it one level farther because of modes. Instead of adding another
check for this specific case that'll probably break in the future too, I
changed it to loop until it reaches the root group and checks that.