Center zoom area around middle of workspace in Push mode
This commit changes the math of determining the viewport
and the mouse position in zoom's Push mouse tracking mode.

The X/Y translation of the viewport is now calculated the
same way as in Centered and Disabled tracking modes. The
old calculations were from Proportional mode, making it
harder to center the zoom area around the mouse.

When zooming in/out, the zoom area will now center around
the visible cursor. This makes it less likely for me to lose
my place when zooming into Plasma.

Signed-off-by: Ritchie Frodomar <alkalinethunder@gmail.com>