scene/surfaceitem: change frame time estimation heuristic
Checking whether or not the surface has stopped rendering based on the last
frame time is error prone, as it may dip below the average quite often.
Instead, this just adds a timeout of 100ms to report the surface as static,
which helps reduce how often videos get wrongly removed from overlay planes.