xwayland: Don't special case selections with no mime types
If you want to clear the clipboard or the primary selection, you need to
disown the corresponding selection, which will trigger x11OfferLost().

The ICCCM spec doesn't say that a selection without mime types should be
handled differently.

Furthermore, both Qt and GTK do nothing special to handle this case,
they treat it as if the current selection had no data. Weird, but it's
easy to make sense of it.

This change removes special casing for selections with no mime types so
the code is more maintainable. As of now, the issue of figuring out
"what does what" still stands, and code like this doesn't help either.