xwayland: Drop SelectionSource::setOffers()
At the moment, mime type handling is somewhat over-engineered and
confusing. The Mimes type is used both to list available mime types and
to cache xcb_atom_t values.

However, the latter is not really worth it. We may need to look up
an atom only when it is time to transfer data, which is a pretty rare
event.

Current code complexity is unjustifiable.