libinput: Add devicesSysNamesByKind for server-side filtering of devices
The DevicesModel used in the Touchscreen KCM queries KWin for all devices
and the fetches the "touchscreen" property on each device to figure out
what touch screens there are.

Rather than issue a bunch of DBus calls like this, let KWin just return
the list of devices it is interested in.