Fix orientation sensor availability
isConnectedToBackend does not mean the backend is available, it means
the sensor is active. It becomes true only when we start the sensor.

It is valid to check it when seeing if a reading is up to date, it is
not correct to theck it to see if the sensor is available.
sensorsForType will do that.

Amends: 854fb6c6978ef8889a506b1e5a73ddc2fe858e1d