plugins/slidingpopups: Adjust repaint region to match animation
When a window moves, we need to repaint the area the window was in on
the next frame.

The current code repaints the final position of a window which works
enough when it's touching a screen edge, but not for cases like the
onscreen keyboard that goes over the panel.

BUG: 510197


(cherry picked from commit a17800d4adbb5c1a1c2357b740deadd8d01bf92c)

a17800d4 plugins/slidingpopups: Adjust repaint region to match animation

Co-authored-by: David Edmundson <kde@davidedmundson.co.uk>