wayland: Drop SeatInterfacePrivate::Keyboard::Focus::destroyConnection
With selections, the surface is not the only relevant part. The client
is relevant too. Currently, the destroyed signal is connected for a
couple of reasons: to reset the surface pointer when the corresponding
surface is destroyed and to clear selection lists. Selection lists don't
necessarily need to be cleared, we can keep them as is until a new
surface is focused. This allows us to simplify selection management code
later.

(cherry picked from commit e200d99760f0eb08c84b9531cbf6f778dc551f9c)