wayland: Reject receive requests from superseded data offers
Currently, if a client loses focus, it can still access the data source.
It is not a critical security issue (the data source has already been
introduced to the client), but it is still a sketchy thing that we do.

The spec also suggests that a data offer should not be accessed after a
leave event.