wayland: Avoid sending a selection when switching between surfaces of the same client
According to the spec, if the keyboard focus moves between two surfaces
that belong to the same client, the selection should not be sent again.

For example, this is relevant with popups. The keyboard focus will move
back and forth between the toplevel and popup surfaces, which belong to
the same client.

offerSelection() and offerPrimarySelection() have been moved to keep
selection related code spatially close for better navigation.