utils/orientationsensor: work around QSensor not having a proper available flag
There's no proper way to check whether or not the sensor is available. The
previous approach of using sensorsForType just reported whether or not it was
connected to iio-sensor-proxy, which is not useful information.
Instead, this just starts the sensor and on the first reading, marks it as
available and disables the sensor again.