backends/wayland: Follow frame callbacks
This fixes nested kwin_wayland running at half of the refresh rate of
the host output.

The issue is that the presentation time protocol gives us page flip
timestamps. If nested kwin_wayland schedules compositing relative to
those timestamps, it may commit the wl_surface after host compositor
composes a frame.

Using frame callbacks results in better refresh rates. It is not
a perfect solution but it is good enough for now.