Guard potential double deletion in InputPanel
InputPanels are marked as destroyed when either the surface is deleted
or the input_panel object is deleted.

If the panel is used within an animation the Window object may stay
alive.

If a surface is destroyed at this point m_panelSurface is potentially
null which is a crash, it's also not ideal to emit signals twice.