backends/wayland: track the commit deadline with frame callbacks
This ensures we get both proper presentation time stamps and can still render
and commit in time for the host compositor to use the frame for the targeted
pageflip timestamp.
To make sure we can still hit the full frame rate in all cases, this also
enables triple buffering for the nested backend.